# openapi.yaml
openapi: 3.0.3
info:
  title: Word Search API
  description: >
    Filters words from dictionary files by available letters, positional hints,
    and word length.
  version: 1.0.0

tags:
  - name: health
    description: Service liveness
  - name: search
    description: Dictionary word search

paths:
  /health:
    get:
      tags: [health]
      summary: Liveness check
      operationId: health
      responses:
        "200":
          description: Service is healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthResponse"

  /search/file:
    post:
      tags: [search]
      summary: Search words of a fixed length
      operationId: searchFile
      description: >
        Returns words of exactly `wordLength` characters that can be formed from
        the available `letters` and satisfy every positional hint.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SearchFileRequest"
      responses:
        "200":
          description: Matching words
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SearchResponse"
        "400":
          description: >
            Invalid request: malformed JSON, a value outside the schema's
            bounds (e.g. unknown `lang`, `wordLength` outside 1–31, hint
            `position` below 1), or both `letters` and `hints` empty.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /search/many:
    post:
      tags: [search]
      summary: Search words across all lengths
      operationId: searchMany
      description: >
        Returns words for every length from 1 up to the number of `letters`,
        ordered longest-first. Useful when the exact word length is unknown.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SearchManyRequest"
      responses:
        "200":
          description: Matching words
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SearchResponse"
        "400":
          description: >
            Invalid request: malformed JSON or a value outside the schema's
            bounds (e.g. unknown `lang`, more than 32 `letters`).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

components:
  schemas:
    HealthResponse:
      type: object
      required:
        - status
      properties:
        status:
          type: string
          description: Always `ok` when the service is up.
          example: ok

    ErrorResponse:
      type: object
      required:
        - error
      properties:
        error:
          type: string
          description: Human-readable reason the request was rejected.
          example: letters and hints cannot both be empty

    Lang:
      type: string
      enum: [fr, en]
      default: fr
      description: Language code — selects the dictionary to search.
      example: fr

    Hint:
      type: object
      description: Positional constraint on a word.
      required:
        - position
      properties:
        position:
          type: integer
          minimum: 1
          maximum: 31
          description: 1-indexed position of the constraint in the word.
          example: 3
        letter:
          type: string
          nullable: true
          description: >
            Expected letter at `position`. Null or absent means no letter
            constraint (only the position matters).
          example: "e"
        excluded:
          type: boolean
          default: false
          description: When true, `letter` must NOT appear at `position`.

    SearchFileRequest:
      type: object
      description: Search for words of one fixed length.
      required:
        - wordLength
      properties:
        lang:
          $ref: "#/components/schemas/Lang"
        wordLength:
          type: integer
          minimum: 1
          maximum: 31
          description: Exact word length to search for.
          example: 5
        letters:
          type: array
          maxItems: 32
          items:
            type: string
          default: []
          description: >
            Available letters, one character per element. Empty means no
            letter-pool constraint (only hints filter).
          example: ["a", "e", "r", "s", "t"]
        hints:
          type: array
          maxItems: 31
          items:
            $ref: "#/components/schemas/Hint"
          default: []
          description: Positional constraints, applied after the letter-pool filter.
        strict:
          type: boolean
          default: false
          description: >
            When true, each element of `letters` may only be used once
            (Scrabble-style).

    SearchManyRequest:
      type: object
      description: Search for words of every length up to the number of letters.
      required:
        - letters
      properties:
        lang:
          $ref: "#/components/schemas/Lang"
        letters:
          type: string
          maxLength: 32
          description: >
            Available letters as a single string. The longest word searched has
            as many characters as `letters`; every shorter length is scanned too.
          example: "artes"
        hints:
          type: array
          maxItems: 31
          items:
            $ref: "#/components/schemas/Hint"
          default: []
          description: Positional constraints, applied to every length scanned.

    SearchResponse:
      type: object
      description: Word search result.
      required:
        - words
        - count
      properties:
        words:
          type: array
          items:
            type: string
          description: Matching words in scan order (longest-first for /search/many).
        count:
          type: integer
          description: Number of matching words (the length of `words`).